Add new data format. IMPORTANT!
To better align with the website, colorCycle is used for both problem
and solution. Problem and solution data are now pointers point to the
color in the color Cycle. This helps with file size as well as easy of
use for the front end.

From now on, both problem and solution are obtained from the same
palleted-reduced run to make sure both dataset align. If
they are ran on different runs of image reduction, the variation in
KMeans can cause misalignment between the same pixels of both dataset.